+++ package.json
@@ -12,6 +12,7 @@
     "@types/node": "^20.12.7",
     "@types/react": "^18.2.78",
     "@types/react-dom": "^18.2.25",
+    "babel-plugin-macros": "^3.1.0",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "firebase": "9.6.6",
@@ -30,7 +31,7 @@
     "serverless-http": "^2.7.0",
     "stripe": "^8.210.0",
     "styled-components": "5.3.3",
-    "typescript": "^5.4.5",
+    "typed-redux-saga": "^1.5.0",
     "update": "^0.7.4",
     "web-vitals": "2.1.4"
   },
@@ -57,5 +58,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/redux-logger": "^3.0.13"
   }
 }
