+++ src/store/cart/cart.selector.ts
@@ -1,7 +1,10 @@
 import { createSelector } from 'reselect';
+
+import { RootState } from '../store';
+
 import { CartState } from './cart.reducer';
 
-const selectCartReducer = (state: any): CartState => state.cart;
+const selectCartReducer = (state: RootState): CartState => state.cart;
 
 export const selectCartItems = createSelector(
   [selectCartReducer],
