+++ src/store/categories/category.selector.ts
@@ -1,9 +1,11 @@
 import { createSelector } from 'reselect';
 
+import { RootState } from '../store';
+
 import { CategoriesState } from './category.reducer';
 import { CategoryMap } from './category.types';
 
-const selectCategoryReducer = (state: { categories: CategoriesState; }): CategoriesState => state.categories;
+const selectCategoryReducer = (state: RootState): CategoriesState => state.categories;
 
 export const selectCategories = createSelector(
   [selectCategoryReducer],
