--- src/store/middleware/logger.js
@@ -1,13 +1,0 @@
-export const loggerMiddleware = (store) => (next) => (action) => {
-  if (!action.type) {
-    return next(action);
-  }
-
-  console.log('type: ', action.type);
-  console.log('payload: ', action.payload);
-  console.log('currentState: ', store.getState());
-
-  next(action);
-
-  console.log('next state: ', store.getState());
-};
