+++ src/store/middleware/logger.ts
@@ -1,0 +1,18 @@
+import { Middleware } from 'redux';
+
+import { RootState } from '../store';
+
+export const loggerMiddleware: Middleware<{}, RootState> =
+  (store) => (next) => (action) => {
+    if (!action.type) {
+      return next(action);
+    }
+
+    console.log('type: ', action.type);
+    console.log('payload: ', action.payload);
+    console.log('currentState: ', store.getState());
+
+    next(action);
+
+    console.log('next state: ', store.getState());
+  };
