+++ src/store/user/user.selector.ts
@@ -1,7 +1,9 @@
 import { createSelector } from "reselect";
 
+import { RootState } from "../store";
+
 import { UserState } from "./user.reducer";
 
-export const selectUserReducer = (state: { user: UserState; }): UserState => state.user;
+export const selectUserReducer = (state: RootState): UserState => state.user;
 
 export const selectCurrentUser = createSelector(selectUserReducer, (user) => user.currentUser);
